LayoutTests/fast/dom/HTMLElement/attr-dir-value-change-auto-parent-expected.txt

PASS getComputedStyle(element).getPropertyValue('direction') is 'rtl'
PASS getComputedStyle(element).getPropertyValue('direction') is 'rtl'
PASS successfullyParsed is true

TEST COMPLETE
مرحبا
